voice_assistant/requirements.txt
# Windows Voice-Controlled Desktop Assistant
# Dependencies

# Core Framework
pydantic>=2.5.0
python-dotenv>=1.0.0

# Screen Capture & OCR
mss>=9.0.1
pillow>=10.2.0
easyocr>=1.7.1
numpy>=1.26.0

# Wake Word Detection
openwakeword>=0.5.1

# Speech Recognition & TTS
SpeechRecognition>=3.10.1
pyttsx3>=2.90

# AI APIs
openai>=1.12.0
httpx>=0.26.0

# System Tray
pystray>=0.19.5

# Audio handling (sounddevice replaces PyAudio - easier Windows install)
sounddevice>=0.4.6
scipy>=1.10.04
